starting to split the big file
