+++ backend/app/db.py
@@ -1,0 +1,13 @@
+import os
+from dotenv import load_dotenv
+from psycopg2 import connect
+
+load_dotenv()
+
+def getConnection():
+    return connect(
+        host = os.environ['DB_HOST'],
+        dbname = os.environ['DB_NAME'],
+        user = os.environ['DB_USER'],
+        password = os.environ['DB_USER_PASSWORD'],
+    )
