+++ backend/app/routes.py
@@ -1,19 +1,10 @@
-import os, json, re
-from dotenv import load_dotenv
+import json, re
 from datetime import datetime
 from flask import abort, jsonify
-from psycopg2 import connect
 from app import app
+from app.db import getConnection
 
-load_dotenv()
-
-def getConnection():
-    return connect(
-        host = os.environ['DB_HOST'],
-        dbname = os.environ['DB_NAME'],
-        user = os.environ['DB_USER'],
-        password = os.environ['DB_USER_PASSWORD'],
-    )
+from app.get_links import get_links
 
 @app.errorhandler(400)
 def request_error(e):
@@ -107,79 +98,7 @@ def get_links_between_two_nodes(from_node_id, to_node_id):
         }
     })
 
-#Function that returns a json with geometries of links between two nodes
-def get_links(from_node_id, to_node_id):
 
-    links_query = '''
-        WITH results as (
-            SELECT *
-            FROM here_gis.get_links_btwn_nodes_22_2(
-                %(node_start)s,
-                %(node_end)s
-            ),
-            UNNEST (links) WITH ORDINALITY AS unnested (link_dir, seq)
-        )
-
-        SELECT 
-            results.link_dir,
-            InitCap(attr.st_name) AS st_name,
-            results.seq,
-            seg_lookup.segment_id,
-            ST_AsGeoJSON(streets.geom) AS geojson,
-            ST_Length( ST_Transform(streets.geom,2952) ) AS length_m
-        FROM results
-        JOIN here.routing_streets_22_2 AS streets USING ( link_dir )
-        JOIN here_gis.streets_att_22_2 AS attr 
-            ON attr.link_id::int = substring(link_dir,'\d+')::int
-        JOIN congestion.network_links_22_2 AS seg_lookup USING ( link_dir )
-        ORDER BY seq;
-        '''
-
-    stname_query = '''
-        SELECT DISTINCT InitCap(st_name) AS st_name
-        FROM here.routing_streets_22_2 AS routing
-        INNER JOIN here_gis.streets_att_22_2 AS streets USING(link_id)
-        WHERE
-            routing.source = %(node)s
-            OR routing.target = %(node)s;
-    '''
-
-    with getConnection() as connection:
-        with connection.cursor() as cursor:
-            cursor.execute(links_query,
-                {"node_start": from_node_id, "node_end": to_node_id}
-            )
-
-            links = [
-                {
-                    'link_dir': link_dir,
-                    'name': st_name,
-                    'sequence': seq,
-                    'segment_id': segment_id,
-                    'geometry': json.loads(geojson),
-                    'length_m': length_m
-                } for link_dir, st_name, seq, segment_id, geojson, length_m in cursor.fetchall()
-            ]
-
-            linknames = set( link['name'] for link in links )
-    
-            cursor.execute(stname_query, {"node": from_node_id})
-            start = set( name for (name,) in cursor.fetchall() ) - linknames
-
-            cursor.execute(stname_query, {"node": to_node_id})
-            end = set( name for (name,) in cursor.fetchall() ) - linknames
-
-            if(len(start) > 0 and len(end) > 0):
-                linkstr = " & ".join(linknames)
-                startstr = " & ".join(start)
-                endstr = " & ".join(end)
-
-                ststr = linkstr + " from " + startstr + " to " + endstr
-            else:
-                ststr = "Along " + " & ".join(linknames)
-
-    connection.close()
-    return links, ststr
 
 
 # test URL /aggregate-travel-times/30310940/30310942/9/12/2020-05-01/2020-06-01/true/2
@@ -231,10 +150,10 @@ def aggregate_travel_times(start_node, end_node, start_time, end_time, start_dat
         #Raise error and return without executing query: dow list does not contain valid characters
         return jsonify({'error': "dow list does not contain valid characters, i.e. [1-7]"})
 
-    holiday_query = ''
+    holiday_subquery = ''
     if include_holidays == 'false':
-        holiday_query = '''AND NOT EXISTS (
-            SELECT 1 FROM ref.holiday WHERE cn.dt = holiday.dt -- excluding holidays
+        holiday_subquery = '''AND NOT EXISTS (
+            SELECT 1 FROM ref.holiday WHERE cn.dt = holiday.dt
         ) '''
     
     agg_tt_query = f''' 
@@ -250,7 +169,7 @@ def aggregate_travel_times(start_node, end_node, start_time, end_time, start_dat
                 AND cn.hr <@ %(time_range)s::numrange
                 AND date_part('isodow', cn.dt)::integer IN %(dow_list)s
                 AND cn.dt <@ %(date_range)s::daterange 
-            {holiday_query}
+            {holiday_subquery}
             GROUP BY
                 cn.dt,
                 cn.hr
